+++ R/package.R
@@ -26,61 +26,9 @@
 NULL
 
 
-#' @importFrom wrapr apply_left
-#' @export
-wrapr::apply_left
-
 #' @importFrom wrapr apply_right
-#' @export
-wrapr::apply_right
+NULL
 
-
-#' @importFrom wrapr mk_tmp_name_source
-#' @export
-wrapr::mk_tmp_name_source
-
-#' @importFrom wrapr map_to_char
-#' @export
-wrapr::map_to_char
-
-#' @importFrom wrapr %.>%
-#' @export
-wrapr::`%.>%`
-
-#' @importFrom wrapr let
-#' @export
-wrapr::let
-
-#' @importFrom wrapr qc
-#' @export
-wrapr::qc
-
-#' @importFrom wrapr qe
-#' @export
-wrapr::qe
-
-#' @importFrom wrapr qae
-#' @export
-wrapr::qae
-
-#' @importFrom wrapr build_frame
-#' @export
-wrapr::build_frame
-
-#' @importFrom wrapr draw_frame
-#' @export
-wrapr::draw_frame
-
-#' @importFrom wrapr qchar_frame
-#' @export
-wrapr::qchar_frame
-
-
-# can collide with data.table := , but we
-# need this for a lot of the pipe examples.
-
-
-#' @importFrom wrapr %:=%
-#' @export
-wrapr::`%:=%`
+#' @importFrom wrapr apply_left
+NULL