+++ NEWS.md
@@ -1,10 +1,11 @@
 
-# rquery 1.4.3 2020/01/27
+# rquery 1.4.3 2020/01/30
 
  * Work on noSuggests issue.
  * Better error message on named argument function.
  * Don't optimize out select_columns as often.
  * Add .(-x) notation to promote string values to names.
+ * Move wrapr to Depends.
 
 # rquery 1.4.2 2020/01/17